Dispatch event after receiving a post-payment request from Postfinance
This allows event listeners to fetch the post-payment parameters and
store for example the used payment method (Credit Card, Postfinance, Twint...)

Note that until now, all post-payment data was stored in the order's data field.
This commit removes this in favor of the event, as this data does not really belong
to the order entity.